+++ src/main/java/com/example/survey/exceptions/ApplicationException.java
@@ -1,0 +1,9 @@
+package com.example.survey.exceptions;
+
+public class ApplicationException extends RuntimeException {
+
+	public ApplicationException(String message) {
+		super(message);
+	}
+
+}
